Log publishOk=false if the SLA has already been exceeded.
If, for some reason, the receipt of a message into PAM is delayed by
more than the threshold, it does not attempt to monitor that item, but
then it isn't captured by the SLA alerts and dashboards. In this case we
can send a message with the negative result to the metrics channel
immediately.

Ignore Eclipse project files.